Revert "EES-1190 Check if footnote linked to another subject for release when deleting."
This reverts commit 0d439823

+++ src/GovUk.Education.ExploreEducationStatistics.Publisher/Functions/PublishReleaseContentImmediateFunction.cs
@@ -54,10 +54,10 @@ namespace GovUk.Education.ExploreEducationStatistics.Publisher.Functions
                 await _contentService.UpdateContentAsync(new[] {message.ReleaseId}, context);
                 await _releaseService.SetPublishedDatesAsync(message.ReleaseId, context.Published);
                 
-//                if (!PublisherUtils.IsDevelopment())
-//                {
+                if (!PublisherUtils.IsDevelopment())
+                {
                     await _releaseService.DeletePreviousVersionsStatisticalData(new[] {message.ReleaseId});
-//                }
+                }
                 
                 await _contentService.DeletePreviousVersionsContent(new[] {message.ReleaseId});
                 await _notificationsService.NotifySubscribersAsync(new[] {message.ReleaseId});
